prem-chat-dolly-v2/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pytest==7.2.2
requests==2.28.2
tqdm==4.65.0
httpx==0.23.3
python-dotenv==1.0.0
llama-cpp-python==0.1.43
